Add files via upload
Commit #1
For the first version of the coding assessment, I made sure that one-to-one mapping exists from string s1 to string s2.
For example: s1 = foo and s2 = bar, the output would return false.

Commit #2
For the second version of the coding assessment, I made sure that one-to-one mapping exists from string s1 to string s2 as well but also wanted to make sure that if s1 was bar and s2 was foo, then I will return true because bar can be mapped to a character in foo.
For example: s1 = bar and s2 = foo, the output would return true.

Commit #3
This is my final solution to the coding assessment and is the same as commit #2, but without a main method. Time complexity is O(n) because only one for loop is being used to compare each character from both strings. Space complexity is O(n) because there are n amount of variables being used.